% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello
Says Hi}
\usage{
hello(a, b, ...)
}
\arguments{
\item{a}{some input}

\item{b}{another input}

\item{...}{unused}
}
\value{
A character sting
}
\description{
Hello
Says Hi
}
\examples{
hello("Annett", "Louis")

}
